simplified wiki link mask getting regex, renamed utils directory into source
